refactor: separated Layout from App
